+++ internal/apps/ohmslaw/ohmslaw.go
@@ -1,12 +1,12 @@
 package ohmslaw
 
 // CalculatecVoltage is return Voltage
-func CalculatecVoltage(current, resistance float32) float32 {
+func CalculatecVoltage(current, resistance float64) float64 {
 	return current * resistance
 }
 
 // CalculateCurrent is return Current
-func CalculateCurrent(voltage, resistance float32) float32 {
+func CalculateCurrent(voltage, resistance float64) float64 {
 	if resistance <= 0 {
 		panic("Invalid resistance.")
 	}
@@ -14,7 +14,7 @@ func CalculateCurrent(voltage, resistance float32) float32 {
 }
 
 // CalculateResistance is return Resistance
-func CalculateResistance(voltage, current float32) float32 {
+func CalculateResistance(voltage, current float64) float64 {
 	if current <= 0 {
 		panic("Invalid current.")
 	}
